refactor: move check url api from ProjectController to RepositoryController
