+++ src/main/java/pvs/app/controller/ProjectController.java
@@ -31,38 +31,6 @@ public class ProjectController {
         this.repositoryService = repositoryService;
     }
 
-    @GetMapping("/repository/github/check")
-    public ResponseEntity<String> checkGitHubURL(@RequestParam("url") String url) {
-        if (repositoryService.checkGithubURL(url)) {
-            return ResponseEntity.status(HttpStatus.OK).body(successMessage);
-        }
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(urlInvalidMessage);
-    }
-
-    @GetMapping("/repository/sonar/check")
-    public ResponseEntity<String> checkSonarURL(@RequestParam("url") String url) {
-        if (repositoryService.checkSonarURL(url)) {
-            return ResponseEntity.status(HttpStatus.OK).body(successMessage);
-        }
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(urlInvalidMessage);
-    }
-
-    @GetMapping("/repository/gitlab/check")
-    public ResponseEntity<String> checkGitLabURL(@RequestParam("url") String url) {
-        if (repositoryService.checkGitLabURL(url)) {
-            return ResponseEntity.status(HttpStatus.OK).body(successMessage);
-        }
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(urlInvalidMessage);
-    }
-
-    @GetMapping("/repository/trello/check")
-    public ResponseEntity<String> checkTrelloURL(@RequestParam("url") String url) {
-        if (repositoryService.checkTrelloURL(url)) {
-            return ResponseEntity.status(HttpStatus.OK).body(successMessage);
-        }
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(urlInvalidMessage);
-    }
-
     @PostMapping("/project")
     public ResponseEntity<String> createProject(@RequestBody CreateProjectDTO createProjectDTO) {
         projectService.create(createProjectDTO);
